refactor(set-frontmatter): migrate imports to _cle-libs
- skills/set-frontmatter/scripts/
  - Update shared imports in set-frontmatter entry point
  - Update libraries, modules, phases, and type imports
  - Replace `_scripts` import paths with `_cle-libs`

- skills/set-frontmatter/scripts/__tests__/
  - Update helper, unit, functional, fixture, and E2E test imports
  - Replace `_scripts` import paths with `_cle-libs`
