+++ skills/set-frontmatter/scripts/__tests__/unit/setfm-count-by-status.unit.spec.ts
@@ -17,9 +17,9 @@ import { describe, it } from '@std/testing/bdd';
 import { _countByStatusForTest as countByStatus } from '../../set-frontmatter.ts';
 
 // ─── Helpers
-import { ChatlogCache } from '../../../../_scripts/classes/ChatlogCache.class.ts';
-import { ChatlogEntry } from '../../../../_scripts/classes/ChatlogEntry.class.ts';
-import { normalizePath } from '../../../../_scripts/libs/path-utils/path-utils.ts';
+import { ChatlogCache } from '../../../../_cle-libs/classes/ChatlogCache.class.ts';
+import { ChatlogEntry } from '../../../../_cle-libs/classes/ChatlogEntry.class.ts';
+import { normalizePath } from '../../../../_cle-libs/libs/path-utils/path-utils.ts';
 // constants
 import { SETFM_CACHE_STATUSES } from '../../types/cache.const.type.ts';
 // types
