+++ skills/set-frontmatter/scripts/libs/__tests__/unit/render-prompt.unit.spec.ts
@@ -14,7 +14,7 @@ import { describe, it } from '@std/testing/bdd';
 import { renderPrompt } from '../../template-utils.ts';
 
 // ─── Helpers
-import { ChatlogError } from '../../../../../_scripts/classes/ChatlogError.class.ts';
+import { ChatlogError } from '../../../../../_cle-libs/classes/ChatlogError.class.ts';
 
 // ─── Tests
