+++ skills/set-frontmatter/scripts/modules/__tests__/unit/setfm-assets-loader.unit.spec.ts
@@ -17,9 +17,9 @@ import { afterEach, beforeEach, describe, it } from '@std/testing/bdd';
 import { loadDics, loadPrompts } from '../../setfm-assets-loader.ts';
 
 // ─── Helpers
-import { GlobalConfig } from '../../../../../_scripts/classes/GlobalConfig.class.ts';
+import { GlobalConfig } from '../../../../../_cle-libs/classes/GlobalConfig.class.ts';
 // constants
-import { DEFAULT_CONFIG_DIR } from '../../../../../_scripts/constants/defaults.constants.ts';
+import { DEFAULT_CONFIG_DIR } from '../../../../../_cle-libs/constants/defaults.constants.ts';
 
 // ─── Internal Helpers
