+++ skills/set-frontmatter/scripts/modules/__tests__/unit/setfm-cache.unit.spec.ts
@@ -14,7 +14,7 @@ import { assertEquals } from '@std/assert';
 import { afterEach, beforeEach, describe, it } from '@std/testing/bdd';
 
 // ─── Test target
-import { ChatlogCache } from '../../../../../_scripts/classes/ChatlogCache.class.ts';
+import { ChatlogCache } from '../../../../../_cle-libs/classes/ChatlogCache.class.ts';
 // types
 import type { SetfmCache } from '../../../types/cache.types.ts';
 // status types
