+++ skills/set-frontmatter/scripts/phases/__tests__/fixtures/phase-type-category.fixtures.spec.ts
@@ -17,9 +17,9 @@ import { beforeEach, describe, it } from '@std/testing/bdd';
 import { phaseTypeAndCategory } from '../../phase-type-category.ts';
 
 // ─── Helpers
-import { ChatlogCache } from '../../../../../_scripts/classes/ChatlogCache.class.ts';
-import { ChatlogEntry } from '../../../../../_scripts/classes/ChatlogEntry.class.ts';
-import { normalizePath } from '../../../../../_scripts/libs/path-utils/path-utils.ts';
+import { ChatlogCache } from '../../../../../_cle-libs/classes/ChatlogCache.class.ts';
+import { ChatlogEntry } from '../../../../../_cle-libs/classes/ChatlogEntry.class.ts';
+import { normalizePath } from '../../../../../_cle-libs/libs/path-utils/path-utils.ts';
 // types
 import type { SetfmCache } from '../../../types/cache.types.ts';
 import type { DicEntry, Dics, Prompts } from '../../../types/dics.types.ts';
