+++ skills/set-frontmatter/scripts/phases/__tests__/unit/phase-frontmatter.unit.spec.ts
@@ -20,10 +20,10 @@ import { needsFrontmatterAi, phaseFrontmatter } from '../../phase-frontmatter.ts
 
 // ─── Helpers
 import { stringify } from '@std/yaml';
-import { ChatlogCache } from '../../../../../_scripts/classes/ChatlogCache.class.ts';
-import { ChatlogEntry } from '../../../../../_scripts/classes/ChatlogEntry.class.ts';
-import { ChatlogError } from '../../../../../_scripts/classes/ChatlogError.class.ts';
-import { logger } from '../../../../../_scripts/libs/io/logger.ts';
+import { ChatlogCache } from '../../../../../_cle-libs/classes/ChatlogCache.class.ts';
+import { ChatlogEntry } from '../../../../../_cle-libs/classes/ChatlogEntry.class.ts';
+import { ChatlogError } from '../../../../../_cle-libs/classes/ChatlogError.class.ts';
+import { logger } from '../../../../../_cle-libs/libs/io/logger.ts';
 // constants
 import { SETFM_CACHE_STATUSES } from '../../../types/cache.const.type.ts';
 // types
