+++ skills/set-frontmatter/scripts/phases/__tests__/unit/phase-write.unit.spec.ts
@@ -17,11 +17,11 @@ import { afterEach, beforeEach, describe, it } from '@std/testing/bdd';
 import { phaseWrite } from '../../phase-write.ts';
 
 // ─── Helpers
-import { makeLoggerStub } from '../../../../../_scripts/__tests__/helpers/logger-stub.ts';
-import { ChatlogCache } from '../../../../../_scripts/classes/ChatlogCache.class.ts';
-import { ChatlogEntry } from '../../../../../_scripts/classes/ChatlogEntry.class.ts';
+import { makeLoggerStub } from '../../../../../_cle-libs/__tests__/helpers/logger-stub.ts';
+import { ChatlogCache } from '../../../../../_cle-libs/classes/ChatlogCache.class.ts';
+import { ChatlogEntry } from '../../../../../_cle-libs/classes/ChatlogEntry.class.ts';
 // types
-import type { LoggerStub } from '../../../../../_scripts/__tests__/helpers/logger-stub.ts';
+import type { LoggerStub } from '../../../../../_cle-libs/__tests__/helpers/logger-stub.ts';
 
 import type { SetfmCache } from '../../../types/cache.types.ts';
 import type { Stats } from '../../../types/phase.types.ts';
